# Updated `romanize()` to improve performance.